Crate simconnect_sys

simconnect-sys

Utilities for interoperating with the SimConnect API

Modules

ffi
Low-level FFI SimConnect APIs

Structs

DataDefinition
A client data definition

EventDefinition
An event definition

NotificationGroupDefinition
A notification group definition

SimConnect
A handle to the SimConnect API

Traits

DataDefinitionGroup
A group of data definitions that can be registered simultaneously

EventType
An event type

NotificationGroup
A set of notification groups that can be registered simultaneously

SimConnectDispatcher
A handler of messages dispatched from SimConnect